+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/transforms/PartiQLSchemaInferencer.kt
@@ -147,7 +147,7 @@ public object PartiQLSchemaInferencer {
                 session = ctx.session,
                 metadata = ctx.metadata,
                 scopingOrder = PlanTyper.ScopingOrder.LEXICAL_THEN_GLOBALS,
-                customFunctions = emptyList(),
+                customFunctionSignatures = emptyList(),
                 problemHandler = ctx.problemHandler
             )
         )